refactor: OntologyEditPage 서브컴포넌트 분해 + 잔여 rgba 토큰화 (A4)
OntologyEditPage.tsx(~2,895줄)에 내장돼 있던 서브컴포넌트를 ui/ 개별
파일로, 순수 함수(formatBuilder*)와 상태 판정 로직을 lib/ 개별 파일로
물리 이동했다 — 기능/props 무변.

- ui/: BuilderCommandStrip · BuilderWriteConfirmBar · BuilderReaderIntentStrip
  · BuilderDetailsDraftCallout · BuilderCanvasEntryRail · BuilderWriteSummary
  · CanvasSkeleton.
- lib/: builder-command-strip-state(상태 판정) · builder-header-action-label
  · format-builder-anchor-labels · is-ontology-kind ·
  build-builder-reader-action-href · builder-draft-agent-packet ·
  builder-verification-packet.
- builder-write-confirm-bar.ts 의 BuilderDraftPreview 참조를 ui →
  builder-draft-agent-packet(lib)로 정정 — 레이어 역참조 제거.
- 분해 후 남은 페이지 인라인 rgba(...) 하드코드(7개 파일에 걸쳐 110건)를
  A2 알파 토큰 37종 우선 재사용해 치환하고, 부족한 opacity/hue(indigo
  a07/13/20/24/30/34/52/60, indigo-line 계열 9종, danger a12/42 +
  text-strong, amber-source, green, overlay/shadow/scrim, panel-overlay)
  만 app/globals.css 에 신설했다 — 시각 결과 무변, 참조만 CSS 변수로 치환.
- 테스트 파일(OntologyEditPage.anchor-rail.test.tsx, BuilderWriteSummary
  .test.tsx) import 경로를 새 위치로 갱신, 토큰화로 바뀐 rgba 리터럴
  assertion 1건을 var(...) 로 정정.

검증: pnpm exec tsc --noEmit / pnpm lint / vitest ontology-edit 28
files·195 tests / i18n 메시지 카탈로그 테스트 모두 그린 / pnpm exec knip
신규 고아 export 0.
